feat: move all queries to one package
