+++ aggsender/query/bridge_query_test.go
@@ -1,4 +1,4 @@
-package bridgequery
+package query
 
 import (
 	"context"
