+++ aggsender/query/l1info_tree_data_query.go
@@ -1,4 +1,4 @@
-package l1infotreequery
+package query
 
 import (
 	"context"
