+++ aggsender/query/l1info_tree_data_query_test.go
@@ -1,4 +1,4 @@
-package l1infotreequery
+package query
 
 import (
 	"context"
